product-gpt-chat/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
google-cloud-firestore==2.13.1
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
requests==2.31.0
pydantic==2.5.0
python-dotenv==1.0.0
cors==1.0.0
functions-framework==3.5.0
mangum==0.17.0
cachecontrol==0.13.1

